feat(plugin): adopt the guest-js convention for tauri-plugin-spindle-project
Closes #32.

`tauri-plugin-spindle-project` was the one plugin in this repo still
invoked through raw `invoke('plugin:spindle-project|...')` calls
scattered across `project-store.ts`, `MenusPage.tsx`, and
`SettingsPage.tsx`, with no `guest-js` package of its own — unlike
`tauri-plugin-display-awareness`, which got the official Tauri
`guest-js`/`dist-js` binding convention in #31.

Added `guest-js/index.ts`, `package.json`
(`tauri-plugin-spindle-project-api`), `rollup.config.js`, and
`tsconfig.json`, mirroring the display-awareness plugin's setup. The
new package exports typed binding functions for every command the
plugin registers (`createProject`, `parseProject`, `serialiseProject`,
`validateProject`, `inspectAsset`, `extractVideoThumbnail`,
`extractImageThumbnail`, `generateBuildPlan`, `executeBuild`,
`cancelBuild`, `autoGenerateMenuNav`, `checkToolchain`,
`exportMenuRenderPreview`, `listAvailableFonts`, `getCacheDir`,
`exportDiagnostics`) plus an `onBuildProgress` wrapper around the
`spindle://build-progress` event the build commands emit.

The project domain model (`SpindleProjectFile`, `Menu`, `Asset`,
`BuildPlan`, etc.) moved from `apps/spindle/src/types/project.ts` into
the plugin's `guest-js`, since those types mirror the Rust structs the
plugin's commands actually serialise — the plugin is now the source of
truth, the same relationship `tauri-plugin-display-awareness` has with
its `DisplayGeometry` type. `apps/spindle/src/types/project.ts` now
re-exports them from `tauri-plugin-spindle-project-api` so existing
imports keep working, and keeps only the app-specific helpers
(`createDefaultProject`, `inferDefaultMenuDisplayAspect`, etc.) that
aren't part of the plugin's own schema.

`generateBuildPlan`, `executeBuild`, and `exportDiagnostics` now take a
single `BuildOptions` argument (`skipSidecar`,
`skipUnsupportedStreams`, `quantizeOverlayPalette`) instead of three
positional booleans, since that was the shape every call site already
threaded through together.

`apps/spindle/package.json` gained
`tauri-plugin-spindle-project-api` as a `workspace:*` dependency and
added it to the existing `predev`/`prebuild`/`pretest`/`pretest:ci`
lifecycle hooks alongside `tauri-plugin-display-awareness-api`, so a
clean checkout builds both plugins' JS bindings before they're needed.
