+++ plugins/tauri-plugin-spindle-project/package.json
@@ -1,0 +1,34 @@
+{
+	"name": "tauri-plugin-spindle-project-api",
+	"version": "0.1.0",
+	"author": "Liminal HQ, Scott Morris",
+	"description": "JS bindings for the spindle-project Tauri plugin.",
+	"type": "module",
+	"types": "./dist-js/index.d.ts",
+	"main": "./dist-js/index.cjs",
+	"module": "./dist-js/index.js",
+	"exports": {
+		"types": "./dist-js/index.d.ts",
+		"import": "./dist-js/index.js",
+		"require": "./dist-js/index.cjs",
+		"default": "./dist-js/index.js"
+	},
+	"files": [
+		"dist-js",
+		"README.md"
+	],
+	"scripts": {
+		"build": "rollup -c",
+		"prepublishOnly": "pnpm build",
+		"pretest": "pnpm build"
+	},
+	"dependencies": {
+		"@tauri-apps/api": "^2.0.0"
+	},
+	"devDependencies": {
+		"@rollup/plugin-typescript": "^12.0.0",
+		"rollup": "^4.9.6",
+		"typescript": "^5.3.3",
+		"tslib": "^2.6.2"
+	}
+}
