+++ plugins/tauri-plugin-spindle-project/rollup.config.js
@@ -1,0 +1,31 @@
+import { readFileSync } from 'node:fs';
+import { dirname, join } from 'node:path';
+import { cwd } from 'node:process';
+import typescript from '@rollup/plugin-typescript';
+
+const pkg = JSON.parse(readFileSync(join(cwd(), 'package.json'), 'utf8'));
+
+export default {
+	input: 'guest-js/index.ts',
+	output: [
+		{
+			file: pkg.exports.import,
+			format: 'esm',
+		},
+		{
+			file: pkg.exports.require,
+			format: 'cjs',
+		},
+	],
+	plugins: [
+		typescript({
+			declaration: true,
+			declarationDir: dirname(pkg.exports.import),
+		}),
+	],
+	external: [
+		/^@tauri-apps\/api/,
+		...Object.keys(pkg.dependencies || {}),
+		...Object.keys(pkg.peerDependencies || {}),
+	],
+};
