+++ plugins/tauri-plugin-spindle-project/tsconfig.json
@@ -1,0 +1,14 @@
+{
+	"compilerOptions": {
+		"target": "es2021",
+		"module": "esnext",
+		"moduleResolution": "bundler",
+		"skipLibCheck": true,
+		"strict": true,
+		"noUnusedLocals": true,
+		"noImplicitAny": true,
+		"noEmit": true
+	},
+	"include": ["guest-js/*.ts"],
+	"exclude": ["dist-js", "node_modules"]
+}
